# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Portfolio API",
    "version": "1.0.0",
    "description": "Modular Express.js API Server for Portfolio Project. Available resources: badges, cvData, giscus, gitHub, gitHubLinks, globalConfig, portfolioData, projects, skills, themes, users, version",
    "contact": {
      "name": "Portfolio API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for Authentication"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/{resource}/stats": {
      "get": {
        "summary": "Get resource statistics",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resource statistics"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}